Coin Problem
Given a set of coin values and an amount of money,
determine the minimum number of coins to make the given
amount of money. The input starts with an integer n,
specifying the number of coin types. Next n integers are the
coin values. The final integer is the amount of money you
have to make. You can assume that the amount will always
be possible to make using the given coin types.